+++ src/main/java/insane96mcp/shieldsplus/world/item/enchantment/AblazeEnchantment.java
@@ -3,6 +3,7 @@ package insane96mcp.shieldsplus.world.item.enchantment;
 import insane96mcp.insanelib.world.enchantments.IEnchantmentTooltip;
 import insane96mcp.insanelib.world.scheduled.ScheduledTasks;
 import insane96mcp.insanelib.world.scheduled.ScheduledTickTask;
+import insane96mcp.shieldsplus.module.BaseFeature;
 import insane96mcp.shieldsplus.setup.SPEnchantments;
 import insane96mcp.shieldsplus.world.item.SPShieldItem;
 import net.minecraft.ChatFormatting;
@@ -17,8 +18,6 @@ import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraftforge.event.entity.living.ShieldBlockEvent;
 
 public class AblazeEnchantment extends Enchantment implements IBlockingEffect, IEnchantmentTooltip {
-
-	public static final int SECONDS_ON_FIRE = 2;
 
 	public AblazeEnchantment() {
 		super(Rarity.UNCOMMON, SPShieldItem.SHIELD, new EquipmentSlot[]{EquipmentSlot.MAINHAND, EquipmentSlot.OFFHAND});
@@ -48,7 +47,7 @@ public class AblazeEnchantment extends Enchantment implements IBlockingEffect, I
 			ScheduledTasks.schedule(new ScheduledTickTask(1) {
 				@Override
 				public void run() {
-					other.setSecondsOnFire(ablaze * AblazeEnchantment.SECONDS_ON_FIRE);
+					other.setSecondsOnFire(ablaze * BaseFeature.enchantmentsAblazeTimeOnFire);
 					if (other instanceof LivingEntity livingEntity) {
 						livingEntity.setLastHurtByMob(attacker);
 						if (livingEntity instanceof Player player)
@@ -60,6 +59,6 @@ public class AblazeEnchantment extends Enchantment implements IBlockingEffect, I
 
 	@Override
 	public Component getTooltip(ItemStack itemStack, int lvl) {
-		return Component.translatable(this.getDescriptionId() + ".tooltip", SECONDS_ON_FIRE * lvl).withStyle(ChatFormatting.DARK_PURPLE);
+		return Component.translatable(this.getDescriptionId() + ".tooltip", BaseFeature.enchantmentsAblazeTimeOnFire * lvl).withStyle(ChatFormatting.DARK_PURPLE);
 	}
 }